begin melding rp2040 and rp2350 APIs
