+++ .gitignore
@@ -37,5 +37,6 @@ test.exe
 test.gba
 test.hex
 test.nro
+test.uf2
 test.wasm
 wasm.wasm
